chore: Domain-Model 생성
