+++ LottoDairy/Domain/Model/Chart/ChartComponents.swift
@@ -1,0 +1,12 @@
+//
+//  ChartComponents.swift
+//  LottoDairy
+//
+//  Created by Sunny on 2023/09/09.
+//
+
+struct ChartComponents {
+
+    var month: Int
+    var account: Double
+}
